Move all components related to Opp, Role, Tool, Team to separate folders
