+++ components/Role/List/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './styles';
 import { ScrollView, View } from 'react-native';
-import RoleCard from '../RoleCard';
+import RoleCard from '../Card';
 
 let RoleList = ({ roles }) =>
   <ScrollView style={styles.roleList}>
