+++ components/Team/Card/styles.js
