+++ components/Tool/List/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './styles';
 import { ScrollView, View } from 'react-native';
-import ToolCard from '../ToolCard';
+import ToolCard from '../Card';
 
 let ToolList = ({ tools }) =>
   <ScrollView style={styles.toolList}>
