+++ components/screens/Discover/index.js
@@ -4,7 +4,7 @@ import { Text } from 'react-native';
 import { graphql } from 'react-apollo';
 import { gql } from 'apollo-boost';
 import ScreenView from '../../ScreenView';
-import OppList from '../../OppList';
+import OppList from '../../Opp/List';
 
 let DiscoverScreen = ({ opps }) =>
   <ScreenView>
