+++ components/screens/Dream/Roles/index.js
@@ -3,7 +3,7 @@ import { compose, mapProps, branch, renderNothing } from 'recompose';
 import { graphql } from 'react-apollo';
 import { gql } from 'apollo-boost';
 import ScreenView from '../../../ScreenView';
-import RoleList from '../../../RoleList';
+import RoleList from '../../../Role/List';
 
 let RolesScreen = ({ roles }) =>
   <ScreenView>
